Entity updates
Added entity file for better toggle behaviour
Added conditional parent to have icon on all parents
Added drawerClose function to close Nav on a tag click
